requirements.txt
requests
python-dotenv
boto3
pydantic
